requirements.txt
fastapi==0.85.0
uvicorn==0.20.0
jinja2==3.1.2
aiofiles==22.1.0
psycopg2-binary==2.9.5
sqlalchemy==1.4.45
python-dotenv==0.21.0
pydantic[email]
passlib[bcrypt]
requests==2.28.1
pytest==7.2.0
httpx==0.23.1
pre-commit==2.20.0
python-jose==3.3.0
python-multipart==0.0.5
python-decouple==3.6
config==0.5.1
